:zap: Change int64_t to int32_t since it matches the specifications of GraphQL

+++ cgql/schema/GraphQLScalar.h
@@ -9,7 +9,7 @@ namespace cgql {
 template<typename T>
 using SerializeFunc = std::function<T(T)>;
 
-typedef int64_t Int;
+typedef int32_t Int;
 typedef std::string_view String;
 
 template<typename T>
